refactor: service 패키지 이름 변경
root/model/common/service -> root/model/common/utils

[skip ci]
