+++ server/edusync/src/main/java/com/codestates/edusync/model/common/utils/VerifyStudygroupCalendarUtils.java
@@ -1,4 +1,4 @@
-package com.codestates.edusync.model.common.service;
+package com.codestates.edusync.model.common.utils;
 
 import com.codestates.edusync.exception.BusinessLogicException;
 import com.codestates.edusync.model.study.plancalendar.entity.TimeSchedule;
