+++ server/edusync/src/main/java/com/codestates/edusync/model/study/postcomment/controller/StudygroupPostCommentController.java
@@ -1,12 +1,12 @@
 package com.codestates.edusync.model.study.postcomment.controller;
 
-import com.codestates.edusync.model.common.service.MemberUtils;
+import com.codestates.edusync.model.common.utils.MemberUtils;
 import com.codestates.edusync.model.member.entity.Member;
 import com.codestates.edusync.model.study.postcomment.entity.StudygroupPostComment;
 import com.codestates.edusync.model.study.postcomment.mapper.StudygroupPostCommentMapper;
 import com.codestates.edusync.model.study.postcomment.service.StudygroupPostCommentService;
 import com.codestates.edusync.model.study.postcomment.dto.StudygroupPostCommentDto;
-import com.codestates.edusync.model.common.service.UriCreator;
+import com.codestates.edusync.model.common.utils.UriCreator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
